// src/models/expense.route.ts
export enum ExpenseRoute  {
    ROOT = "/expenses",
    CREATE = "/",
    UPDATE = "/:id",
    DELETE = "/:id",
    GET_EXPENSES = "/",
    CREATE_PATH = "/expenses",
    UPDATE_PATH = "/expenses/:id",
    DELETE_PATH = "/expenses/:id",
    GET_EXPENSES_PATH = "/expenses",
    GET_EXPENSE = "/:id"
}